Improve wait cc running test
