+++ pkg/cmd/run/clusterclaim/exec.go
@@ -46,7 +46,7 @@ func (o *Options) run() (err error) {
 			return err
 		}
 	}
-	err = clusterpoolhost.RunClusterClaims(o.ClusterClaims, o.SkipSchedule, o.CMFlags.DryRun, o.outputFile)
+	err = clusterpoolhost.RunClusterClaims(o.ClusterClaims, o.SkipSchedule, o.Timeout, o.CMFlags.DryRun, o.outputFile)
 	if err != nil {
 		return err
 	}
